Split down command structure
