+++ lib/commands/utility_commands.rb
@@ -2,7 +2,6 @@
 
 require 'thor'
 require_relative '../utilities/utilities'
-require_relative '../utilities/open_ai'
 
 class UtilityCommands < Thor
   desc 'path [PATH]', 'Sets the default path for scaffolding'
@@ -18,15 +17,11 @@ class UtilityCommands < Thor
     Utilities.send("#{type}_path=", default_path)
   end
 
-  map '-p' => 'path'
-
   desc 'url [URL]', 'Sets the default url for a project'
 
   def url(default_url)
     Utilities.url = default_url
   end
-
-  map '-u' => 'url'
 
   desc 'browser [BROWSER]', 'Sets the default browser for a project'
   option :opts,
@@ -39,8 +34,6 @@ class UtilityCommands < Thor
     browser_options(options[:opts]) if options[:opts] || options[:delete]
   end
 
-  map '-b' => 'browser'
-
   desc 'browser_options [OPTIONS]', 'Sets the browser options for the project'
   option :delete,
          type: :boolean, required: false, desc: 'This will delete your browser options', aliases: '-d'
@@ -49,8 +42,6 @@ class UtilityCommands < Thor
     Utilities.browser_options = opts unless opts.empty?
     Utilities.delete_browser_options if options[:delete]
   end
-
-  map '-bo' => 'browser_options'
 
   desc 'raid', 'It runs all the tests in a project'
   option :parallel,
@@ -66,8 +57,6 @@ class UtilityCommands < Thor
     end
   end
 
-  map '-r' => 'raid'
-
   desc 'config', 'Creates a new config file'
   option :path,
          type: :string, required: false, desc: 'The path where your config file will be created', aliases: '-p'
@@ -80,46 +69,10 @@ class UtilityCommands < Thor
     Utilities.platform = platform
   end
 
-  map '-pl' => 'platform'
-
-  desc 'download_builds [BUILD_TYPE]', 'It downloads the example builds for appium projects'
-  def download_builds(build_type)
+  desc 'builds [BUILD_TYPE]', 'It downloads the example builds for appium projects'
+  def builds(build_type)
     raise 'Please select one of the following build types: android, ios, both' unless %w[android ios both].include?(build_type)
 
     Utilities.download_builds build_type
   end
-
-  map '-d' => 'download_builds'
-
-  desc 'version', 'It shows the version of Ruby Raider you are currently using'
-  def version
-    spec = Gem::Specification.load('ruby_raider.gemspec')
-    version = spec.version
-    puts "The Ruby Raider version is #{version}, Happy testing!"
-  end
-
-  map '-v' => 'version'
-
-  desc 'open_ai [REQUEST]', 'Uses open AI to create a file or generate output'
-  option :path,
-         type: :string, required: false, desc: 'The path where your file will be created', aliases: '-p'
-  option :edit,
-         type: :string, required: false, desc: 'Path to the file you want to edit', aliases: '-e'
-
-  def open_ai(request, path = nil)
-    path ||= options[:path]
-    if options[:edit]
-      pp 'Editing File...'
-      OpenAi.edit_file(options[:edit], request)
-      pp "File #{options[:edit]} edited"
-    elsif path
-      pp 'Generating File...'
-      OpenAi.create_file(path, request)
-      pp "File created in #{path}"
-    else
-      puts OpenAi.output(request)
-    end
-  end
-
-  map '-oa' => 'open_ai'
 end
