+++ lib/generators/templates/common/read_me.tt
@@ -54,7 +54,9 @@ Ruby Raider is a generator and scaffolding gem to make UI test automation easier
 
 ***In order to run the Appium tests, download the example [app](https://github.com/saucelabs/my-demo-app-rn).***
 
-***In order to run the visual tests with applitools, you need to create an account and get your api key, you can read more [here](https://applitools.com/docs/topics/overview/obtain-api-key.html#:~:text=If%20you%20already%20have%20an,Your%20key%20will%20be%20displayed.).***
+***In order to run the visual tests with applitools, you need to create an account and get your api key, you can read
+more [here](https://applitools.com/docs/topics/overview/obtain-api-key.html#:~:text=If%20you%20already%20have%20an,Your%20key%20will%20be%20displayed.)
+.***
 
 ***To use the open ai integration you need to set up the OPENAI_ACCESS_TOKEN environment variable and
 you can also set the optional OPENAI_ORGANIZATION_ID if you have an organization***
@@ -77,52 +79,38 @@ gem install ruby_raider
 raider new [name_of_project]
 ```
 
-Then a TUI/CLI will appear where the configuration of which frameworks you want to be generated/scaffolded can be selected.
+Then a TUI/CLI will appear where the configuration of which frameworks you want to be generated/scaffolded can be
+selected.
 
 Select the ones you will like to work with.
 
-### Ruby raider provides the following list of commands
+### Ruby raider provides the following list of basic commands
 
 ###### Anything between square brackets([...]) is where your imput goes
 
 ```ruby
-Commands:
-  raider browser [BROWSER]          # Sets the default browser for a project
-
-  raider browser_options [OPTIONS]  # Sets the browser options for the project
-
-  raider feature [FEATURE_NAME]     # Creates a new feature
-
-  raider help [COMMAND]             # Describes available commands or one specific command
-
-  raider helper [HELPER_NAME]       # Creates a new helper
-
-  raider new [PROJECT_NAME]         # Creates a new framework based on settings picked
-
-  raider open_ai [REQUEST]          # Uses open AI to create a file or generate output
-
-  raider page [PAGE_NAME]           # Creates a new page object
-
-  raider path [PATH]                # Sets the default path for scaffolding
-
-  raider platform [PLATFORM]        # Sets the default platform for a cross-platform project
-
-  raider raid                       # Runs all the tests in a project
-
-  raider scaffold [SCAFFOLD_NAME]   # Generates everything needed to start automating
-
-  raider spec [SPEC_NAME]           # Creates a new spec
-
-  raider url [URL]                  # Sets the default url for a project
-
-  raider version                    # It shows the version of Ruby Raider you are currently using
+Commands :
+  raider generate # Provides access to all the generators commands
+raider help [COMMAND] # Describe available commands or one specific command
+raider new [PROJECT_NAME] # Creates a new framework based on settings picked
+raider open_ai # Provides access to all the open ai commands
+raider utility # Provides access to all the utility commands
+raider version # It shows the version of Ruby Raider you are currently using
 ```
+
+All the basic commands have their corresponding shortcut:
+
+* g for generate
+* n for new
+* o for open_ai
+* u for utility
+* v for version
 
 Ruby raider also supports scaffolding:
 
-* To create a new page object you do: ```raider page [PAGE_NAME]```
-* To create a new spec you do: ```raider spec [SPEC_NAME]```
-* To create a new feature you do: ```raider feature [FEATURE_NAME]```
+* To create a new page object you do: ```raider g page [PAGE_NAME]```
+* To create a new spec you do: ```raider g spec [SPEC_NAME]```
+* To create a new feature you do: ```raider g feature [FEATURE_NAME]```
 
 It's possible to add the option --path or -p if you want to specify where to create your features, pages, helpers and
 specs.
@@ -130,22 +118,30 @@ specs.
 If you want to set the default path for the creation of your features, helpers and specs:
 
 ```ruby
-raider path [PATH_NAME] --feature or -f
-raider path [PATH_NAME] --spec or -s
-raider path [PATH_NAME] --helper or -h
+raider u path [PATH_NAME] - -feature or -f
+raider u path [PATH_NAME] - -spec or -s
+raider u path [PATH_NAME] - -helper or -h
 ```
 
 If you don't specify an option, path will assume you want to change the default path for pages.
 
 ### Open AI Commands
 
-If you want to set the default path for the creation of your features, helpers and specs:
-
 ```ruby
 # Will print the result of the request on the terminal
-raider open_ai [REQUEST]
+raider o make [REQUEST]
 # Will create a file with the result of your request as content
-raider open_ai [REQUEST] --path or -p [PATH]
+raider o make [REQUEST] - -path or -p [PATH]
 # Will input the content of the chosen file into open ai and will edit it based on the result
-raider path [PATH_NAME] --edit or -e [FILE_PATH]
+raider o make [PATH_NAME] - -edit or -e [FILE_PATH]
+# Creates a cucumber file and uses it to input into open ai and create a steps file
+# The prompt is not required because it uses a based prompt
+raider o cucumber [NAME] - -prompt or -p [PROMPT]
+# Creates a cucumber step definitions file based on an scenario file
+raider open_ai steps [NAME]
+Options :
+  -p, [--path = PATH] # The path where your steps will be created
+  -pr, [--prompt = PROMPT] # This will create the selected steps based on your prompt using open ai
+  -i, [--input = INPUT] # It uses a file as input to create the steps
+
 ```
