+++ lib/open_ai/open_ai.rb
